Link pages using sample data
