+++ components/card/App.js
@@ -1,14 +1,21 @@
 import Screen from "components/card/Screen.js";
 import { join } from "path";
+import Link from "next/link";
 
 const AppCard = ({ app }) => (
   <div>
     <div className="flex mt-10">
       <div>
-        <img className="app-logo" src={app.logo} />
+        <Link href={`/apps/${app.slug}`}>
+          <a>
+            <img className="app-logo" src={app.logo} />
+          </a>
+        </Link>
       </div>
       <div className="mx-3 mt-2 text-sm font-bold">
-        <h3 className=""> {app.name} </h3>
+        <h3 className="">
+          <Link href={`/apps/${app.slug}`}>{app.name}</Link>{" "}
+        </h3>
       </div>
     </div>
