+++ components/sections/HeaderView.js
@@ -1,18 +1,25 @@
 import React from "react";
-const HeaderView = () => (
+import Link from "next/link";
+
+const HeaderView = ({ app }) => (
   <div className="">
     <div className="mt-5">
-      <a className="back-button rounded-full bg-gray-100 px-5 py-3 font-bold text-xs"><img src="/img/arrow.png" className="inline-block mr-2"/>Back</a>
+      <Link href="/">
+      <a className="back-button rounded-full bg-gray-100 px-5 py-3 font-bold text-xs">
+        <img src="/img/arrow.png" className="inline-block mr-2" />
+        Back
+      </a>
+      </Link>
     </div>
     <div className="text-center mt-7">
       <div className="flex justify-center">
-        <img src="/img/bitpay.png" />
+        <img src={app.logo} />
       </div>
       <p className="text-sm mt-4 font-extrabold leading-6 tracking-tighter">
-        BitPay - Buy Crypto
+        {app.name}
       </p>
       <span className="text-sm text-gray-500 mt-1 tracking-tighter leading-6">
-        A Secure Bitcoin Wallet
+        {app.description}
       </span>
     </div>
     <div className="flex">
