+++ lib/sample.js
@@ -1,0 +1,39 @@
+const apps = [
+  {
+    name: "Argent - DeFi in a tap",
+    logo: "/img/1.png",
+    type: "mobile",
+    slug: "argent-de-fi-in-a-tap",
+    screens: [
+      { image: "/img/2.png" },
+      { image: "/img/4.png" },
+      { image: "/img/5.png" },
+      { image: "/img/6.png" },
+      { image: "/img/7.png" },
+      { image: "/img/8.png" },
+    ],
+  },
+  {
+    name: "Atomic Wallet",
+    logo: "/img/logo2.png",
+    type: "mobile",
+    slug: "atomic-wallet",
+    screens: [
+      { image: "/img/11.png" },
+      { image: "/img/12.png" },
+      { image: "/img/13.png" },
+      { image: "/img/14.png" },
+      { image: "/img/15.png" },
+      { image: "/img/16.png" },
+    ],
+  },
+  {
+    name: "Desktop app",
+    logo: "/img/1.png",
+    type: "desktop",
+    slug: "desktop-app",
+    screens: [{ image: "/img/3.png" }, { image: "/img/3.png" }],
+  },
+];
+
+export default apps;
