+++ pages/index.tsx
@@ -2,47 +2,12 @@ import Layout from "../components/Layout";
 import React from "react";
 import AppCard from "../components/card/App";
 import Header from "../components/sections/Header";
-
+import SampleData from '../lib/sample';
 
 
 
 const IndexPage = () => {
-  const apps = [
-    {
-      name: "Argent - DeFi in a tap",
-      logo: "/img/1.png",
-      type: "mobile",
-      screens: [
-        { image: "/img/2.png" },
-        { image: "/img/4.png" },
-        { image: "/img/5.png" },
-        { image: "/img/6.png" },
-        { image: "/img/7.png" },
-        { image: "/img/8.png" },
-      ],
-    },
-
-    {
-      name: "Atomic Wallet",
-      logo: "/img/logo2.png",
-      type: "mobile",
-      screens: [
-        { image: "/img/11.png" },
-        { image: "/img/12.png" },
-        { image: "/img/13.png" },
-        { image: "/img/14.png" },
-        { image: "/img/15.png" },
-        { image: "/img/16.png" },
-      ],
-    },
-
-    {
-      name: "Desktop app",
-      logo: "/img/1.png",
-      type: "desktop",
-      screens: [{ image: "/img/3.png" }, { image: "/img/3.png" }],
-    },
-  ];
+  const apps = SampleData;
   return (
     <Layout title="Home | Next.js + TypeScript Example">
       <main className="w-11/12 mx-auto">
